refactor(goesr): rename netcdf_geodetic to  geodetic
