+++ src/goesdl/goesr/geodetic.py
